WIP loguru testing
