+++ xclim/testing/tests/test_logging.py
@@ -1,56 +1,72 @@
+import logging
 import sys
+from pathlib import Path
 
+import pytest
 from loguru import logger
+
+from xclim import enable_synced_logger
+from xclim.testing._utils import _logging_examples  # noqa
 
 
 class TestLoggingFuncs:
-    @staticmethod
-    def _logging_examples() -> None:
-        """Testing module"""
-        logger.trace("0")
-        logger.debug("1")
-        logger.info("2")
-        logger.success("2.5")
-        logger.warning("3")
-        logger.error("4")
-        logger.critical("5")
-
     def test_logging_configuration(self, capsys):
-        logger.add(sys.stderr, level="WARNING")
-        logger.add(sys.stdout, level="INFO")
+        id1 = logger.add(sys.stdout, level="ERROR")
+        id2 = logger.add(sys.stderr, level="INFO")
 
-        self._logging_examples()  # noqa
-
-        captured = capsys.readouterr()
-        assert "WARNING" in captured.err
-        assert "INFO" in captured.out
-
-        # teardown
-        logger.remove()
-
-    def test_disabled_enabled_logging(self, capsys):
-        logger.add(sys.stderr, level="WARNING")
-        logger.add(sys.stdout, level="INFO")
-
-        # disable xclim logging
+        logger.enable("xclim")
+        _logging_examples()  # noqa
         logger.disable("xclim")
 
-        self._logging_examples()  # noqa
+        captured = capsys.readouterr()
+        assert "ERROR" in captured.out
+        assert "WARNING" not in captured.out
+        assert "WARNING" in captured.err
+
+        # teardown
+        logger.remove(id1)
+        logger.remove(id2)
+
+    def test_disabled_enabled_logging(self, capsys):
+        id1 = logger.add(sys.stderr, level="WARNING")
+        id2 = logger.add(sys.stdout, level="CRITICAL")
+
+        _logging_examples()  # noqa
 
         captured = capsys.readouterr()
         assert "WARNING" not in captured.err
-        assert "INFO" not in captured.out
+        assert "CRITICAL" not in captured.out
 
         # enable xclim logging
         logger.enable("xclim")
-
-        self._logging_examples()  # noqa
-
-        captured = capsys.readouterr()
-        assert "INFO" not in captured.err
-        assert "WARNING" in captured.err
-        assert "INFO" in captured.out
+        _logging_examples()  # noqa
         logger.disable("xclim")
 
+        captured = capsys.readouterr()
+        assert "WARNING" in captured.err
+        assert "WARNING" not in captured.out
+        assert "CRITICAL" in captured.out
+
         # teardown
-        logger.remove()
+        logger.remove(id1)
+        logger.remove(id2)
+
+    @pytest.xfail("This test needs love.")
+    def test_standard_logging_configuration(self, caplog):
+        logger.enable("xclim")
+        enable_synced_logger()
+
+        _logging_examples()  # noqa
+
+        assert ("xclim.testing._utils", 40, "4") in caplog.record_tuples
+
+    def test_file_logger_enable(self, tmpdir):
+        test_log = Path(tmpdir).joinpath("xclim_test.log")
+        logger.add(test_log, level=logging.WARNING)
+
+        logger.enable("xclim")
+        _logging_examples()
+        logger.disable("xclim")
+
+        assert "INFO" not in test_log.read_text()
+        assert "CRITICAL" in test_log.read_text()
